Rename files from QnA Advisor to Customer Advisor
