+++ app-modules/ai/src/Filament/Resources/CustomerAdvisors/CustomerAdvisorResource.php
@@ -34,19 +34,19 @@
 </COPYRIGHT>
 */
 
-namespace AdvisingApp\Ai\Filament\Resources\QnaAdvisors;
+namespace AdvisingApp\Ai\Filament\Resources\CustomerAdvisors;
 
-use AdvisingApp\Ai\Filament\Resources\QnaAdvisors\Pages\CreateQnaAdvisor;
-use AdvisingApp\Ai\Filament\Resources\QnaAdvisors\Pages\EditQnaAdvisor;
-use AdvisingApp\Ai\Filament\Resources\QnaAdvisors\Pages\EditQnaAdvisorLinks;
-use AdvisingApp\Ai\Filament\Resources\QnaAdvisors\Pages\ListQnaAdvisors;
-use AdvisingApp\Ai\Filament\Resources\QnaAdvisors\Pages\ManageCategories;
-use AdvisingApp\Ai\Filament\Resources\QnaAdvisors\Pages\ManageQnaAdditionalKnowledge;
-use AdvisingApp\Ai\Filament\Resources\QnaAdvisors\Pages\ManageQnaAdvisorResourceHub;
-use AdvisingApp\Ai\Filament\Resources\QnaAdvisors\Pages\ManageQnaQuestions;
-use AdvisingApp\Ai\Filament\Resources\QnaAdvisors\Pages\PreviewQnaAdvisor;
-use AdvisingApp\Ai\Filament\Resources\QnaAdvisors\Pages\QnaAdvisorEmbed;
-use AdvisingApp\Ai\Filament\Resources\QnaAdvisors\Pages\ViewQnaAdvisor;
+use AdvisingApp\Ai\Filament\Resources\CustomerAdvisors\Pages\CreateCustomerAdvisor;
+use AdvisingApp\Ai\Filament\Resources\CustomerAdvisors\Pages\EditCustomerAdvisor;
+use AdvisingApp\Ai\Filament\Resources\CustomerAdvisors\Pages\EditCustomerAdvisorLinks;
+use AdvisingApp\Ai\Filament\Resources\CustomerAdvisors\Pages\ListCustomerAdvisors;
+use AdvisingApp\Ai\Filament\Resources\CustomerAdvisors\Pages\ManageCategories;
+use AdvisingApp\Ai\Filament\Resources\CustomerAdvisors\Pages\ManageCustomerAdditionalKnowledge;
+use AdvisingApp\Ai\Filament\Resources\CustomerAdvisors\Pages\ManageCustomerAdvisorResourceHub;
+use AdvisingApp\Ai\Filament\Resources\CustomerAdvisors\Pages\ManageCustomerQuestions;
+use AdvisingApp\Ai\Filament\Resources\CustomerAdvisors\Pages\PreviewCustomerAdvisor;
+use AdvisingApp\Ai\Filament\Resources\CustomerAdvisors\Pages\CustomerAdvisorEmbed;
+use AdvisingApp\Ai\Filament\Resources\CustomerAdvisors\Pages\ViewCustomerAdvisor;
 use AdvisingApp\Ai\Models\CustomerAdvisor;
 use App\Features\RenameQnaAdvisorsFeature;
 use Filament\Pages\Page;
@@ -54,7 +54,7 @@ use Filament\Resources\Resource;
 use Override;
 use UnitEnum;
 
-class QnaAdvisorResource extends Resource
+class CustomerAdvisorResource extends Resource
 {
     protected static ?string $model = CustomerAdvisor::class;
 
@@ -75,32 +75,32 @@ class QnaAdvisorResource extends Resource
     public static function getPages(): array
     {
         return [
-            'index' => ListQnaAdvisors::route('/'),
-            'create' => CreateQnaAdvisor::route('/create'),
-            'view' => ViewQnaAdvisor::route('/{record}'),
-            'edit' => EditQnaAdvisor::route('/{record}/edit'),
-            'edit-websites' => EditQnaAdvisorLinks::route('/{record}/websites'),
+            'index' => ListCustomerAdvisors::route('/'),
+            'create' => CreateCustomerAdvisor::route('/create'),
+            'view' => ViewCustomerAdvisor::route('/{record}'),
+            'edit' => EditCustomerAdvisor::route('/{record}/edit'),
+            'edit-websites' => EditCustomerAdvisorLinks::route('/{record}/websites'),
             'manage-categories' => ManageCategories::route('/{record}/categories'),
-            'manage-questions' => ManageQnaQuestions::route('/{record}/questions'),
-            'manage-additional-knowledge' => ManageQnaAdditionalKnowledge::route('/{record}/additional-knowledge'),
-            'manage-resource-hub' => ManageQnaAdvisorResourceHub::route('/{record}/resource-hub'),
-            'preview' => PreviewQnaAdvisor::route('/{record}/preview'),
-            'embed' => QnaAdvisorEmbed::route('/{record}/embed'),
+            'manage-questions' => ManageCustomerQuestions::route('/{record}/questions'),
+            'manage-additional-knowledge' => ManageCustomerAdditionalKnowledge::route('/{record}/additional-knowledge'),
+            'manage-resource-hub' => ManageCustomerAdvisorResourceHub::route('/{record}/resource-hub'),
+            'preview' => PreviewCustomerAdvisor::route('/{record}/preview'),
+            'embed' => CustomerAdvisorEmbed::route('/{record}/embed'),
         ];
     }
 
     public static function getRecordSubNavigation(Page $page): array
     {
         return $page->generateNavigationItems([
-            ViewQnaAdvisor::class,
-            EditQnaAdvisor::class,
+            ViewCustomerAdvisor::class,
+            EditCustomerAdvisor::class,
             ManageCategories::class,
-            ManageQnaQuestions::class,
-            ManageQnaAdditionalKnowledge::class,
-            ManageQnaAdvisorResourceHub::class,
-            EditQnaAdvisorLinks::class,
-            PreviewQnaAdvisor::class,
-            QnaAdvisorEmbed::class,
+            ManageCustomerQuestions::class,
+            ManageCustomerAdditionalKnowledge::class,
+            ManageCustomerAdvisorResourceHub::class,
+            EditCustomerAdvisorLinks::class,
+            PreviewCustomerAdvisor::class,
+            CustomerAdvisorEmbed::class,
         ]);
     }
 }
