+++ app-modules/ai/src/Http/Controllers/CustomerAdvisors/AuthenticationRefreshController.php
@@ -59,7 +59,7 @@ class AuthenticationRefreshController
         ])
             ->withCookie(
                 Cookie::make(
-                    name: 'advising_app_qna_advisor_refresh_token',
+                    name: 'advising_app_customer_advisor_refresh_token',
                     value: $tokens['refresh_token']->plainTextToken,
                     minutes: 60 * 24 * 3, // 3 days
                     secure: true,
