+++ app-modules/authorization/src/Enums/PermissionDescription.php
@@ -52,6 +52,8 @@ enum PermissionDescription: string implements HasLabel
     case CaseAssignment = 'This permission group enables the display and management of the Case Assignment feature in the quaternary navigation group View Case under a Student or Prospect. Permission options include create, view, update, and delete case assignments.';
     case CaseUpdate = 'This permission group enables the display and management of the Case Update feature in the quaternary navigation group View Case under a Student or Prospect. Permission options include create, view, update, and delete case updates.';
     case Concern = 'This permission group enables the display and management of the Concern feature in the tertiary navigation group View Student or View Prospect. Permission options include create, view, update, and delete concerns.';
+    case CustomerAdvisor = 'This permission group enables the display and management of the Customer Advisor feature in the primary navigation group Chatbots. Permission options include create, view, update, and delete customer advisors.';
+    case CustomerAdvisorEmbed = 'This permission group enables the display and management of the Customer Advisor Embed feature in the tertiary navigation group View Customer Advisor. Permission options include view customer advisor embeds.';
     case Division = 'This permission group enables the display and management of the Division feature in the primary navigation group User Management. Permission options include create, view, update, and delete divisions.';
     case Engagement = 'This permission group enables the display and management of the Sent Items feature in the secondary navigation group Unified Inbox. Permission options include create, view, update, and delete engagements.';
     case EngagementResponse = 'This permission group enables the display and management of the Inbox feature in the secondary navigation group Unified Inbox. Permission options include create, view, update, and delete engagement responses.';
@@ -69,8 +71,6 @@ enum PermissionDescription: string implements HasLabel
     case Project = 'This permission group enables the display and management of the Project feature in the primary navigation group CRM. Permission options include create, view, update, and delete projects.';
     case Prompt = 'This permission group enables the display and management of the Prompt Library feature in the primary navigation group Enterprise AI. Permission options include create, view, update, and delete prompts.';
     case Prospect = 'This permission group enables the display and management of the Prospect feature in the primary navigation group CRM. Permission options include create, view, update, delete, and import prospects.';
-    case QnAAdvisor = 'This permission group enables the display and management of the Customer Advisor feature in the primary navigation group Chatbots. Permission options include create, view, update, and delete customer advisors.';
-    case QnAAdvisorEmbed = 'This permission group enables the display and management of the Customer Advisor Embed feature in the tertiary navigation group View Customer Advisor. Permission options include view customer advisor embeds.';
     case RealtimeChat = 'This permission group enables the display and management of the Realtime Chat feature in the primary navigation group Premium Features. Permission options include view realtime chats.';
     case RecordSync = 'This permission group enables the display and management of the Sync History feature in the primary navigation group Data and Analytics. Permission options include view record syncs.';
     case Report = 'This permission group enables the display and management of the Custom Reports feature in the primary navigation group Data and Analytics. Permission options include create, view, update, and delete custom reports.';
